pyproject.toml
[tool.poetry]
authors = ["hegasz <henry.gasztowtt@cs.ox.ac.uk>", "benbind", "ezhang7423"]
description = ""
name = "large_legislative_models"
readme = "README.md"
version = "1.0.0"

[tool.poetry.dependencies]
eztils = "^0.4.94"
python = "^3.11,<3.12"
python-dotenv = "^1.0.0"
torch = "^2.3.1"
wandb = "^0.16.2"
#dm-meltingpot = "^2.2.2"

cloudpickle = "^3.0.0"
dm-env = "^1.6"
gymnasium = "^0.29.1"
huggingface-hub = "^0.23.4"
loguru = "^0.7.2"
matplotlib = "^3.9.0"
ml-collections = "^0.1.1"
nvitop = "^1.3.2"
pettingzoo = "^1.24.3"
reactivex = "^4.0.4"
stable-baselines3 = "^2.3.2"
supersuit = "^3.9.2"
tensordict = "^0.4.0"
torchvision = "^0.18.1"
chex = "^0.1.86"
lupa = "^2.2"
art = "^6.2"
dm-meltingpot = {path = "meltingpot", develop = true}
pyav = "^12.2.0"
ffmpeg = "^1.4"
av = "^12.2.0"
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"large_legislative_models" = "large_legislative_models.main:main"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
pytest = "^7.4.4"
pyupgrade = "^3.15.0"

####################################
# Tools #
####################################
ipdb = "^0.13.13"
ipykernel = "^6.29.4"
ipython = "^8.20.0"
pyflyby = "^1.9.2"

[tool.black]
# https://github.com/psf/black
color = true
line-length = 120
target-version = ["py38"]

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
line_length = 88
py_version = 38

color_output = true
include_trailing_comma = true
indent = 4
known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
multi_line_output = 3
profile = "black"
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
color_output = true
pretty = true
python_version = "3.11"
show_traceback = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
norecursedirs = ["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
